Split immediate agent action tests
