--- Tests/QuillCodeAgentTests/AgentImmediateActionTests.swift
@@ -1,711 +1,0 @@
-import XCTest
-import QuillCodeCore
-import QuillCodeTools
-@testable import QuillCodeAgent
-
-final class AgentImmediateActionTests: XCTestCase {
-    func testRunWhoamiExecutesImmediately() async throws {
-        let root = try makeTempDirectory()
-        let result = try await AgentRunner().send("run whoami", in: ChatThread(mode: .auto), workspaceRoot: root)
-        XCTAssertEqual(result.toolResults.count, 1)
-        XCTAssertTrue(result.toolResults[0].ok, result.toolResults[0].error ?? "")
-        XCTAssertTrue(result.thread.messages.last?.content.hasPrefix("You are `") == true)
-    }
-
-    func testWhoamiQuestionExecutesImmediatelyWithoutConfirmationLoop() async throws {
-        let root = try makeTempDirectory()
-        let result = try await AgentRunner().send("whoami?", in: ChatThread(mode: .auto), workspaceRoot: root)
-
-        XCTAssertEqual(result.toolResults.count, 1)
-        XCTAssertTrue(result.toolResults[0].ok, result.toolResults[0].error ?? "")
-        XCTAssertTrue(result.thread.messages.last?.content.hasPrefix("You are `") == true)
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("I'll run") })
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("No shell command was specified") })
-
-        XCTAssertEqual(try queuedShellCommand(in: result), "whoami")
-    }
-
-    func testDiskUsageQuestionExecutesImmediately() async throws {
-        let root = try makeTempDirectory()
-        let result = try await AgentRunner().send("How much hd?", in: ChatThread(mode: .auto), workspaceRoot: root)
-
-        XCTAssertEqual(result.toolResults.count, 1)
-        let toolResult = try XCTUnwrap(result.toolResults.first)
-        XCTAssertTrue(toolResult.ok, toolResult.error ?? "")
-        XCTAssertEqual(try queuedShellCommand(in: result), "df -h / /Quill 2>/dev/null || df -h /")
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("I'll check") })
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("No shell command was specified") })
-    }
-
-    func testOpenClawDiscoveryExecutesImmediately() async throws {
-        let root = try makeTempDirectory()
-        let result = try await AgentRunner().send("Do you have openclaw?", in: ChatThread(mode: .auto), workspaceRoot: root)
-
-        XCTAssertEqual(result.toolResults.count, 1)
-        let toolResult = try XCTUnwrap(result.toolResults.first)
-        XCTAssertTrue(toolResult.ok, toolResult.error ?? "")
-        XCTAssertEqual(try queuedShellCommand(in: result), "command -v openclaw || which openclaw || echo 'not found'")
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("I'll check") })
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("No shell command was specified") })
-    }
-
-    func testOpenClawDiscoveryDoesNotDependOnProviderKnowledge() async throws {
-        let root = try makeTempDirectory()
-        let runner = AgentRunner(llm: FailingLLMClient(), enablesImmediateActionPreflight: true)
-        let result = try await runner.send("Do you have openclaw?", in: ChatThread(mode: .auto), workspaceRoot: root)
-
-        XCTAssertEqual(result.toolResults.count, 1)
-        XCTAssertTrue(result.toolResults[0].ok, result.toolResults[0].error ?? "")
-        XCTAssertEqual(try queuedShellCommand(in: result), "command -v openclaw || which openclaw || echo 'not found'")
-        XCTAssertEqual(result.thread.messages.last?.content, "openclaw is not installed or is not on PATH.")
-    }
-
-    func testDiskUsageQuestionDoesNotDependOnProviderKnowledge() async throws {
-        let root = try makeTempDirectory()
-        let runner = AgentRunner(llm: FailingLLMClient(), enablesImmediateActionPreflight: true)
-        let result = try await runner.send("How much hd?", in: ChatThread(mode: .auto), workspaceRoot: root)
-
-        XCTAssertEqual(result.toolResults.count, 1)
-        XCTAssertTrue(result.toolResults[0].ok, result.toolResults[0].error ?? "")
-        XCTAssertEqual(try queuedShellCommand(in: result), "df -h / /Quill 2>/dev/null || df -h /")
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("I'll check") })
-    }
-
-    func testListFilesQuestionExecutesImmediatelyWithoutProviderRoundTrip() async throws {
-        let root = try makeTempDirectory()
-        try "alpha\n".write(to: root.appendingPathComponent("alpha.txt"), atomically: true, encoding: .utf8)
-        try "beta\n".write(to: root.appendingPathComponent("beta.txt"), atomically: true, encoding: .utf8)
-        let runner = AgentRunner(llm: FailingLLMClient(), enablesImmediateActionPreflight: true)
-
-        let result = try await runner.send(
-            "Can you list the files here?",
-            in: ChatThread(mode: .auto),
-            workspaceRoot: root
-        )
-
-        XCTAssertEqual(result.toolResults.count, 1)
-        XCTAssertTrue(result.toolResults[0].ok, result.toolResults[0].error ?? "")
-        let list = try queuedFileList(in: result)
-        XCTAssertEqual(list.path, ".")
-        XCTAssertFalse(list.includeHidden)
-        XCTAssertTrue(result.thread.messages.last?.content.contains("alpha.txt") == true)
-        XCTAssertTrue(result.thread.messages.last?.content.contains("beta.txt") == true)
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("I'll list") })
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("No shell command was specified") })
-    }
-
-    func testListFilesCanScopeToDirectoryAndIncludeHidden() async throws {
-        let root = try makeTempDirectory()
-        try FileManager.default.createDirectory(at: root.appendingPathComponent("Sources"), withIntermediateDirectories: true)
-        try "visible\n".write(to: root.appendingPathComponent("Sources/App.swift"), atomically: true, encoding: .utf8)
-        try "hidden\n".write(to: root.appendingPathComponent("Sources/.secret"), atomically: true, encoding: .utf8)
-        let runner = AgentRunner(llm: FailingLLMClient(), enablesImmediateActionPreflight: true)
-
-        let result = try await runner.send(
-            "Can you list all files in Sources?",
-            in: ChatThread(mode: .auto),
-            workspaceRoot: root
-        )
-
-        XCTAssertEqual(result.toolResults.count, 1)
-        XCTAssertTrue(result.toolResults[0].ok, result.toolResults[0].error ?? "")
-        let list = try queuedFileList(in: result)
-        XCTAssertEqual(list.path, "Sources")
-        XCTAssertTrue(list.includeHidden)
-        XCTAssertTrue(result.thread.messages.last?.content.contains("Sources/.secret") == true)
-        XCTAssertTrue(result.thread.messages.last?.content.contains("Sources/App.swift") == true)
-    }
-
-    func testCurrentDirectoryQuestionExecutesImmediatelyWithoutProviderRoundTrip() async throws {
-        let root = try makeTempDirectory()
-        let runner = AgentRunner(llm: FailingLLMClient(), enablesImmediateActionPreflight: true)
-
-        let result = try await runner.send(
-            "Can you show me the current directory?",
-            in: ChatThread(mode: .auto),
-            workspaceRoot: root
-        )
-
-        XCTAssertEqual(result.toolResults.count, 1)
-        XCTAssertTrue(result.toolResults[0].ok, result.toolResults[0].error ?? "")
-        XCTAssertEqual(try queuedShellCommand(in: result), "pwd")
-        XCTAssertTrue(result.thread.messages.last?.content.contains(root.path) == true)
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("I'll show") })
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("No shell command was specified") })
-    }
-
-    func testPoliteGitStatusUsesStructuredGitStatusBeforeGenericShellRecovery() async throws {
-        let root = try makeTempDirectory()
-        try initializeGitRepo(at: root)
-        try "working tree\n".write(to: root.appendingPathComponent("status.txt"), atomically: true, encoding: .utf8)
-        let runner = AgentRunner(llm: FailingLLMClient(), enablesImmediateActionPreflight: true)
-
-        let result = try await runner.send(
-            "Please check git status.",
-            in: ChatThread(mode: .auto),
-            workspaceRoot: root
-        )
-
-        XCTAssertEqual(result.toolResults.count, 1)
-        XCTAssertTrue(result.toolResults[0].ok, result.toolResults[0].error ?? "")
-        let call = try queuedToolCall(in: result)
-        XCTAssertEqual(call.name, ToolDefinition.gitStatus.name)
-        XCTAssertEqual(call.argumentsJSON, "{}")
-        XCTAssertTrue(result.thread.messages.last?.content.contains("Git status:") == true)
-        XCTAssertTrue(result.thread.messages.last?.content.contains("status.txt") == true)
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("I'll check") })
-    }
-
-    func testDownloadDomainExecutesImmediatelyWithWorkspaceBoundedPath() async throws {
-        let root = try makeTempDirectory()
-        let runner = AgentRunner(toolExecutionOverride: { call, _ in
-            guard call.name == ToolDefinition.shellRun.name else { return nil }
-            return ToolResult(
-                ok: true,
-                stdout: "-rw-r--r--  1 mock  staff  42K downloads/linkedin.com.html\n",
-                stderr: "",
-                exitCode: 0
-            )
-        })
-
-        let result = try await runner.send("Can you download LinkedIn.com?", in: ChatThread(mode: .auto), workspaceRoot: root)
-
-        XCTAssertEqual(result.toolResults.count, 1)
-        let toolResult = try XCTUnwrap(result.toolResults.first)
-        XCTAssertTrue(toolResult.ok, toolResult.error ?? "")
-        XCTAssertEqual(
-            try queuedShellCommand(in: result),
-            "mkdir -p 'downloads' && curl -L --fail --silent --show-error --output 'downloads/linkedin.com.html' 'https://LinkedIn.com' && ls -lh 'downloads/linkedin.com.html'"
-        )
-        XCTAssertEqual(result.thread.messages.last?.content, "Downloaded to `downloads/linkedin.com.html`.")
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("I'll download") })
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("No shell command was specified") })
-    }
-
-    func testDownloadURLIntoExplicitPathUsesRequestedPath() async throws {
-        let root = try makeTempDirectory()
-        let runner = AgentRunner(toolExecutionOverride: { call, _ in
-            guard call.name == ToolDefinition.shellRun.name else { return nil }
-            return ToolResult(
-                ok: true,
-                stdout: "-rw-r--r--  1 mock  staff  559B downloads/example.html\n",
-                stderr: "",
-                exitCode: 0
-            )
-        })
-
-        let result = try await runner.send(
-            "Download https://example.com into `downloads/example.html` in this workspace.",
-            in: ChatThread(mode: .auto),
-            workspaceRoot: root
-        )
-
-        XCTAssertEqual(result.toolResults.count, 1)
-        let toolResult = try XCTUnwrap(result.toolResults.first)
-        XCTAssertTrue(toolResult.ok, toolResult.error ?? "")
-        XCTAssertEqual(
-            try queuedShellCommand(in: result),
-            "mkdir -p 'downloads' && curl -L --fail --silent --show-error --output 'downloads/example.html' 'https://example.com' && ls -lh 'downloads/example.html'"
-        )
-        XCTAssertEqual(result.thread.messages.last?.content, "Downloaded to `downloads/example.html`.")
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("I'll download") })
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("No shell command was specified") })
-    }
-
-    func testGitStatusExecutesImmediatelyWithoutProviderRoundTrip() async throws {
-        let root = try makeTempDirectory()
-        try initializeGitRepo(at: root)
-        try "working tree\n".write(to: root.appendingPathComponent("status.txt"), atomically: true, encoding: .utf8)
-        let runner = AgentRunner(llm: FailingLLMClient(), enablesImmediateActionPreflight: true)
-
-        let result = try await runner.send("git status", in: ChatThread(mode: .auto), workspaceRoot: root)
-
-        XCTAssertEqual(result.toolResults.count, 1)
-        XCTAssertTrue(result.toolResults[0].ok, result.toolResults[0].error ?? "")
-        let call = try queuedToolCall(in: result)
-        XCTAssertEqual(call.name, ToolDefinition.gitStatus.name)
-        XCTAssertEqual(call.argumentsJSON, "{}")
-        XCTAssertTrue(result.thread.messages.last?.content.contains("Git status:") == true)
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("I'll check") })
-    }
-
-    func testWhatChangedUsesGitDiffImmediatelyWithoutProviderRoundTrip() async throws {
-        let root = try makeTempDirectory()
-        try initializeGitRepo(at: root)
-        let path = root.appendingPathComponent("tracked.txt")
-        try "before\n".write(to: path, atomically: true, encoding: .utf8)
-        XCTAssertTrue(GitToolExecutor().stage(cwd: root, path: "tracked.txt").ok)
-        XCTAssertTrue(GitToolExecutor().commit(cwd: root, message: "Add tracked file").ok)
-        try "after\n".write(to: path, atomically: true, encoding: .utf8)
-        let runner = AgentRunner(llm: FailingLLMClient(), enablesImmediateActionPreflight: true)
-
-        let result = try await runner.send("what changed?", in: ChatThread(mode: .auto), workspaceRoot: root)
-
-        XCTAssertEqual(result.toolResults.count, 1)
-        XCTAssertTrue(result.toolResults[0].ok, result.toolResults[0].error ?? "")
-        let call = try queuedToolCall(in: result)
-        XCTAssertEqual(call.name, ToolDefinition.gitDiff.name)
-        XCTAssertEqual(call.argumentsJSON, "{}")
-        XCTAssertTrue(result.thread.messages.last?.content.contains("Git diff:") == true)
-        XCTAssertTrue(result.thread.messages.last?.content.contains("tracked.txt") == true)
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("I'll review") })
-    }
-
-    func testBacktickCommandDoesNotBecomeEmptyToolCall() async throws {
-        let root = try makeTempDirectory()
-        let result = try await AgentRunner().send("Run `pwd`", in: ChatThread(mode: .auto), workspaceRoot: root)
-        XCTAssertEqual(result.toolResults.count, 1)
-        XCTAssertTrue(result.toolResults[0].ok, result.toolResults[0].error ?? "")
-    }
-
-    func testPoliteDoItNowBacktickCommandExecutesImmediately() async throws {
-        let root = try makeTempDirectory()
-        let runner = AgentRunner(llm: FailingLLMClient(), enablesImmediateActionPreflight: true)
-
-        let result = try await runner.send(
-            "Please run `printf quillcode_now_smoke` now and report the output.",
-            in: ChatThread(mode: .auto),
-            workspaceRoot: root
-        )
-
-        XCTAssertEqual(result.toolResults.count, 1)
-        XCTAssertTrue(result.toolResults[0].ok, result.toolResults[0].error ?? "")
-        XCTAssertEqual(try queuedShellCommand(in: result), "printf quillcode_now_smoke")
-        XCTAssertEqual(result.thread.messages.last?.content, "Output:\nquillcode_now_smoke")
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("I'll run") })
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("No shell command was specified") })
-    }
-
-    func testPoliteBareCommandExecutesImmediatelyWithoutProviderRoundTrip() async throws {
-        let root = try makeTempDirectory()
-        let runner = AgentRunner(llm: FailingLLMClient(), enablesImmediateActionPreflight: true)
-
-        let result = try await runner.send(
-            "Can you run printf quillcode_polite_smoke?",
-            in: ChatThread(mode: .auto),
-            workspaceRoot: root
-        )
-
-        XCTAssertEqual(result.toolResults.count, 1)
-        XCTAssertTrue(result.toolResults[0].ok, result.toolResults[0].error ?? "")
-        XCTAssertEqual(try queuedShellCommand(in: result), "printf quillcode_polite_smoke")
-        XCTAssertEqual(result.thread.messages.last?.content, "Output:\nquillcode_polite_smoke")
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("I'll run") })
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("No shell command was specified") })
-    }
-
-    func testIncidentalBareCommandMentionStillUsesModel() async throws {
-        let root = try makeTempDirectory()
-        let runner = AgentRunner(llm: FixedSayLLMClient(message: "Mention noted."), enablesImmediateActionPreflight: true)
-
-        let result = try await runner.send(
-            "The docs say run ls after setup.",
-            in: ChatThread(mode: .auto),
-            workspaceRoot: root
-        )
-
-        XCTAssertEqual(result.toolResults.count, 0)
-        XCTAssertEqual(result.thread.messages.last?.content, "Mention noted.")
-    }
-
-    func testNegatedWhoamiDoesNotPreflight() async throws {
-        let root = try makeTempDirectory()
-        let runner = AgentRunner(llm: FixedSayLLMClient(message: "I will not run it."), enablesImmediateActionPreflight: true)
-
-        let result = try await runner.send(
-            "Do not run whoami.",
-            in: ChatThread(mode: .auto),
-            workspaceRoot: root
-        )
-
-        XCTAssertEqual(result.toolResults.count, 0)
-        XCTAssertEqual(result.thread.messages.last?.content, "I will not run it.")
-    }
-
-    func testNegatedDiagnosticDoesNotPreflight() async throws {
-        let root = try makeTempDirectory()
-        let runner = AgentRunner(llm: FixedSayLLMClient(message: "I will leave disk usage alone."), enablesImmediateActionPreflight: true)
-
-        let result = try await runner.send(
-            "Please don't check disk usage.",
-            in: ChatThread(mode: .auto),
-            workspaceRoot: root
-        )
-
-        XCTAssertEqual(result.toolResults.count, 0)
-        XCTAssertEqual(result.thread.messages.last?.content, "I will leave disk usage alone.")
-    }
-
-    func testNegatedOpenClawDiscoveryDoesNotPreflight() async throws {
-        let root = try makeTempDirectory()
-        let runner = AgentRunner(llm: FixedSayLLMClient(message: "I will not inspect OpenClaw."), enablesImmediateActionPreflight: true)
-
-        let result = try await runner.send(
-            "Do not check openclaw.",
-            in: ChatThread(mode: .auto),
-            workspaceRoot: root
-        )
-
-        XCTAssertEqual(result.toolResults.count, 0)
-        XCTAssertEqual(result.thread.messages.last?.content, "I will not inspect OpenClaw.")
-    }
-
-    func testNegatedDownloadDoesNotPreflight() async throws {
-        let root = try makeTempDirectory()
-        let runner = AgentRunner(llm: FixedSayLLMClient(message: "I will not download it."), enablesImmediateActionPreflight: true)
-
-        let result = try await runner.send(
-            "Don't download https://example.com.",
-            in: ChatThread(mode: .auto),
-            workspaceRoot: root
-        )
-
-        XCTAssertEqual(result.toolResults.count, 0)
-        XCTAssertEqual(result.thread.messages.last?.content, "I will not download it.")
-    }
-
-    func testNegatedFileWriteDoesNotPreflight() async throws {
-        let root = try makeTempDirectory()
-        let runner = AgentRunner(llm: FixedSayLLMClient(message: "I will not create a file."), enablesImmediateActionPreflight: true)
-
-        let result = try await runner.send(
-            "Do not write a file that says hello world.",
-            in: ChatThread(mode: .auto),
-            workspaceRoot: root
-        )
-
-        XCTAssertEqual(result.toolResults.count, 0)
-        XCTAssertEqual(result.thread.messages.last?.content, "I will not create a file.")
-        XCTAssertFalse(FileManager.default.fileExists(atPath: root.appendingPathComponent("hello.txt").path))
-    }
-
-    func testDefaultMockRespectsOnlyNegatedActionRequest() async throws {
-        let root = try makeTempDirectory()
-
-        let result = try await AgentRunner().send(
-            "Do not write a file that says hello world.",
-            in: ChatThread(mode: .auto),
-            workspaceRoot: root
-        )
-
-        XCTAssertEqual(result.toolResults.count, 0)
-        XCTAssertEqual(result.thread.messages.last?.content, "Okay, I won't take that action.")
-        XCTAssertFalse(FileManager.default.fileExists(atPath: root.appendingPathComponent("hello.txt").path))
-    }
-
-    func testAffirmedSecondClauseStillExecutesImmediately() async throws {
-        let root = try makeTempDirectory()
-        let runner = AgentRunner(llm: FailingLLMClient(), enablesImmediateActionPreflight: true)
-
-        let result = try await runner.send(
-            "Do not run whoami; run pwd.",
-            in: ChatThread(mode: .auto),
-            workspaceRoot: root
-        )
-
-        XCTAssertEqual(result.toolResults.count, 1)
-        XCTAssertTrue(result.toolResults[0].ok, result.toolResults[0].error ?? "")
-        XCTAssertEqual(try queuedShellCommand(in: result), "pwd")
-        XCTAssertTrue(result.thread.messages.last?.content.contains(root.path) == true)
-    }
-
-    func testDefaultMockUsesAffirmedActionAfterNegatedClause() async throws {
-        let root = try makeTempDirectory()
-
-        let result = try await AgentRunner().send(
-            "Do not run whoami. Thanks. Run pwd.",
-            in: ChatThread(mode: .auto),
-            workspaceRoot: root
-        )
-
-        XCTAssertEqual(result.toolResults.count, 1)
-        XCTAssertTrue(result.toolResults[0].ok, result.toolResults[0].error ?? "")
-        XCTAssertEqual(try queuedShellCommand(in: result), "pwd")
-        XCTAssertTrue(result.thread.messages.last?.content.contains(root.path) == true)
-    }
-
-    func testMakeHelloWorldFileExecutesImmediately() async throws {
-        let root = try makeTempDirectory()
-        let result = try await AgentRunner().send(
-            "Can you write a file that says hello world",
-            in: ChatThread(mode: .auto),
-            workspaceRoot: root
-        )
-        XCTAssertEqual(result.toolResults.count, 1)
-        XCTAssertTrue(result.toolResults[0].ok, result.toolResults[0].error ?? "")
-        let text = try String(contentsOf: root.appendingPathComponent("hello.txt"), encoding: .utf8)
-        XCTAssertEqual(text, "hello world\n")
-        XCTAssertEqual(result.thread.messages.last?.content, "Wrote `hello.txt`.")
-    }
-
-    func testFollowUpReadIgnoresPreviousToolFeedback() async throws {
-        let root = try makeTempDirectory()
-        let runner = AgentRunner()
-        let writeResult = try await runner.send(
-            "Can you write a file that says hello world",
-            in: ChatThread(mode: .auto),
-            workspaceRoot: root
-        )
-
-        let readResult = try await runner.send(
-            "Read `hello.txt` and tell me its exact content",
-            in: writeResult.thread,
-            workspaceRoot: root
-        )
-
-        XCTAssertEqual(try queuedFileRead(in: readResult), "hello.txt")
-        XCTAssertEqual(readResult.thread.messages.last?.content, "Contents of `hello.txt`:\nhello world")
-        XCTAssertEqual(
-            readResult.thread.messages.filter { $0.role == .assistant }.map(\.content),
-            ["Wrote `hello.txt`.", "Contents of `hello.txt`:\nhello world"]
-        )
-    }
-
-    func testNamedFileWriteExecutesImmediatelyWithoutProviderRoundTrip() async throws {
-        let root = try makeTempDirectory()
-        let runner = AgentRunner(llm: FailingLLMClient(), enablesImmediateActionPreflight: true)
-
-        let result = try await runner.send(
-            "Create a file named notes/todo.txt that says \"buy milk\"",
-            in: ChatThread(mode: .auto),
-            workspaceRoot: root
-        )
-
-        XCTAssertEqual(result.toolResults.count, 1)
-        XCTAssertTrue(result.toolResults[0].ok, result.toolResults[0].error ?? "")
-        XCTAssertEqual(
-            try String(contentsOf: root.appendingPathComponent("notes/todo.txt"), encoding: .utf8),
-            "buy milk\n"
-        )
-        let write = try queuedFileWrite(in: result)
-        XCTAssertEqual(write.path, "notes/todo.txt")
-        XCTAssertEqual(write.content, "buy milk\n")
-    }
-
-    func testBacktickFileReadExecutesImmediatelyWithoutProviderRoundTrip() async throws {
-        let root = try makeTempDirectory()
-        try "hello world\n".write(to: root.appendingPathComponent("hello.txt"), atomically: true, encoding: .utf8)
-        let runner = AgentRunner(llm: FailingLLMClient(), enablesImmediateActionPreflight: true)
-
-        let result = try await runner.send(
-            "Read `hello.txt` and tell me its exact content",
-            in: ChatThread(mode: .auto),
-            workspaceRoot: root
-        )
-
-        XCTAssertEqual(result.toolResults.count, 1)
-        XCTAssertTrue(result.toolResults[0].ok, result.toolResults[0].error ?? "")
-        XCTAssertEqual(try queuedFileRead(in: result), "hello.txt")
-        XCTAssertEqual(result.thread.messages.last?.content, "Contents of `hello.txt`:\nhello world")
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("I'll read") })
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("No shell command was specified") })
-    }
-
-    func testNaturalFileReadExecutesImmediatelyWithStructuredFileTool() async throws {
-        let root = try makeTempDirectory()
-        try "# QuillCode\n".write(to: root.appendingPathComponent("README.md"), atomically: true, encoding: .utf8)
-        let runner = AgentRunner(llm: FailingLLMClient(), enablesImmediateActionPreflight: true)
-
-        let result = try await runner.send(
-            "What is in README.md?",
-            in: ChatThread(mode: .auto),
-            workspaceRoot: root
-        )
-
-        XCTAssertEqual(result.toolResults.count, 1)
-        XCTAssertTrue(result.toolResults[0].ok, result.toolResults[0].error ?? "")
-        XCTAssertEqual(try queuedFileRead(in: result), "README.md")
-        XCTAssertEqual(result.thread.messages.last?.content, "Contents of `README.md`:\n# QuillCode")
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("I'll read") })
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("I will read") })
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("No shell command was specified") })
-    }
-
-    func testNaturalFileSearchExecutesImmediatelyWithStructuredFileTool() async throws {
-        let root = try makeTempDirectory()
-        let sources = root.appendingPathComponent("Sources")
-        try FileManager.default.createDirectory(at: sources, withIntermediateDirectories: true)
-        try "struct AgentRunner {}\n".write(
-            to: sources.appendingPathComponent("Agent.swift"),
-            atomically: true,
-            encoding: .utf8
-        )
-        let runner = AgentRunner(llm: FailingLLMClient(), enablesImmediateActionPreflight: true)
-
-        let result = try await runner.send(
-            "Find AgentRunner",
-            in: ChatThread(mode: .auto),
-            workspaceRoot: root
-        )
-
-        XCTAssertEqual(result.toolResults.count, 1)
-        XCTAssertTrue(result.toolResults[0].ok, result.toolResults[0].error ?? "")
-        let search = try queuedFileSearch(in: result)
-        XCTAssertEqual(search.query, "AgentRunner")
-        XCTAssertNil(search.path)
-        XCTAssertTrue(result.thread.messages.last?.content.contains("Found 1 match for `AgentRunner`:") == true)
-        XCTAssertTrue(result.thread.messages.last?.content.contains("`Sources/Agent.swift:1`") == true)
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("I'll search") })
-        XCTAssertFalse(result.thread.messages.contains { $0.content.contains("No shell command was specified") })
-    }
-
-    func testOpenClawFindRequestKeepsAvailabilityDiagnosticPriority() async throws {
-        let root = try makeTempDirectory()
-        let runner = AgentRunner(llm: FailingLLMClient(), enablesImmediateActionPreflight: true)
-
-        let result = try await runner.send(
-            "find openclaw",
-            in: ChatThread(mode: .auto),
-            workspaceRoot: root
-        )
-
-        XCTAssertEqual(result.toolResults.count, 1)
-        XCTAssertEqual(try queuedShellCommand(in: result), "command -v openclaw || which openclaw || echo 'not found'")
-    }
-
-    func testFileWriteWithQuotedContentDefaultsToNotePath() async throws {
-        let root = try makeTempDirectory()
-        let runner = AgentRunner(llm: FailingLLMClient(), enablesImmediateActionPreflight: true)
-
-        let result = try await runner.send(
-            "Make a file with content `ship the first build`",
-            in: ChatThread(mode: .auto),
-            workspaceRoot: root
-        )
-
-        XCTAssertEqual(result.toolResults.count, 1)
-        XCTAssertTrue(result.toolResults[0].ok, result.toolResults[0].error ?? "")
-        XCTAssertEqual(
-            try String(contentsOf: root.appendingPathComponent("note.txt"), encoding: .utf8),
-            "ship the first build\n"
-        )
-        let write = try queuedFileWrite(in: result)
-        XCTAssertEqual(write.path, "note.txt")
-        XCTAssertEqual(write.content, "ship the first build\n")
-    }
-
-    func testAmbiguousMakeFileRequestStillUsesModel() async throws {
-        let root = try makeTempDirectory()
-        let runner = AgentRunner(llm: FixedSayLLMClient(message: "What should the file contain?"), enablesImmediateActionPreflight: true)
-
-        let result = try await runner.send("Make a file", in: ChatThread(mode: .auto), workspaceRoot: root)
-
-        XCTAssertEqual(result.toolResults.count, 0)
-        XCTAssertEqual(result.thread.messages.last?.content, "What should the file contain?")
-    }
-
-    func testCommitChangesExecutesImmediately() async throws {
-        let root = try makeTempDirectory()
-        try initializeGitRepo(at: root)
-        try "hello\n".write(to: root.appendingPathComponent("hello.txt"), atomically: true, encoding: .utf8)
-        XCTAssertTrue(GitToolExecutor().stage(cwd: root, path: "hello.txt").ok)
-
-        let result = try await AgentRunner().send(
-            "commit these changes with message Add hello file",
-            in: ChatThread(mode: .auto),
-            workspaceRoot: root
-        )
-
-        XCTAssertEqual(result.toolResults.count, 1)
-        XCTAssertTrue(result.toolResults[0].ok, result.toolResults[0].error ?? "")
-        let log = ShellToolExecutor().run(.init(command: "git log -1 --pretty=%s", cwd: root))
-        XCTAssertEqual(log.stdout.trimmingCharacters(in: .whitespacesAndNewlines), "Add hello file")
-    }
-
-    func testPushCurrentBranchExecutesImmediately() async throws {
-        let parent = try makeTempDirectory()
-        let root = parent.appendingPathComponent("repo")
-        let remote = parent.appendingPathComponent("remote.git")
-        try FileManager.default.createDirectory(at: root, withIntermediateDirectories: true)
-        try initializeGitRepo(at: root)
-        XCTAssertTrue(ShellToolExecutor().run(.init(command: "git init --bare '\(remote.path)'", cwd: parent)).ok)
-        XCTAssertTrue(ShellToolExecutor().run(.init(command: "git remote add origin '\(remote.path)'", cwd: root)).ok)
-        try "hello\n".write(to: root.appendingPathComponent("hello.txt"), atomically: true, encoding: .utf8)
-        XCTAssertTrue(GitToolExecutor().stage(cwd: root, path: "hello.txt").ok)
-        XCTAssertTrue(GitToolExecutor().commit(cwd: root, message: "Add hello").ok)
-
-        let result = try await AgentRunner().send(
-            "push this branch",
-            in: ChatThread(mode: .auto),
-            workspaceRoot: root
-        )
-
-        XCTAssertEqual(result.toolResults.count, 1)
-        XCTAssertTrue(result.toolResults[0].ok, result.toolResults[0].error ?? "")
-        XCTAssertEqual(result.thread.events.filter { $0.summary.contains("host.git.push") }.count, 3)
-    }
-
-    private func queuedShellCommand(in result: AgentRunResult) throws -> String {
-        let call = try queuedToolCall(in: result)
-        let arguments = try ToolArguments(call.argumentsJSON)
-        XCTAssertEqual(call.name, ToolDefinition.shellRun.name)
-        return try arguments.requiredString("cmd")
-    }
-
-    private func queuedFileWrite(in result: AgentRunResult) throws -> (path: String, content: String) {
-        let call = try queuedToolCall(in: result)
-        let arguments = try ToolArguments(call.argumentsJSON)
-        XCTAssertEqual(call.name, ToolDefinition.fileWrite.name)
-        return (
-            try arguments.requiredString("path"),
-            try arguments.requiredString("content")
-        )
-    }
-
-    private func queuedFileRead(in result: AgentRunResult) throws -> String {
-        let queued = try XCTUnwrap(result.thread.events.last { $0.kind == .toolQueued })
-        let payloadJSON = try XCTUnwrap(queued.payloadJSON)
-        let call = try JSONDecoder().decode(ToolCall.self, from: Data(payloadJSON.utf8))
-        let arguments = try ToolArguments(call.argumentsJSON)
-        XCTAssertEqual(call.name, ToolDefinition.fileRead.name)
-        return try arguments.requiredString("path")
-    }
-
-    private func queuedFileList(in result: AgentRunResult) throws -> (path: String, includeHidden: Bool) {
-        let queued = try XCTUnwrap(result.thread.events.last { $0.kind == .toolQueued })
-        let payloadJSON = try XCTUnwrap(queued.payloadJSON)
-        let call = try JSONDecoder().decode(ToolCall.self, from: Data(payloadJSON.utf8))
-        let arguments = try ToolArguments(call.argumentsJSON)
-        XCTAssertEqual(call.name, ToolDefinition.fileList.name)
-        return (
-            arguments.string("path") ?? ".",
-            arguments.bool("includeHidden") ?? false
-        )
-    }
-
-    private func queuedFileSearch(in result: AgentRunResult) throws -> (query: String, path: String?) {
-        let queued = try XCTUnwrap(result.thread.events.last { $0.kind == .toolQueued })
-        let payloadJSON = try XCTUnwrap(queued.payloadJSON)
-        let call = try JSONDecoder().decode(ToolCall.self, from: Data(payloadJSON.utf8))
-        let arguments = try ToolArguments(call.argumentsJSON)
-        XCTAssertEqual(call.name, ToolDefinition.fileSearch.name)
-        return (
-            try arguments.requiredString("query"),
-            arguments.string("path")
-        )
-    }
-
-    private func queuedToolCall(in result: AgentRunResult) throws -> ToolCall {
-        let queued = try XCTUnwrap(result.thread.events.first { $0.kind == .toolQueued })
-        let payloadJSON = try XCTUnwrap(queued.payloadJSON)
-        return try JSONDecoder().decode(ToolCall.self, from: Data(payloadJSON.utf8))
-    }
-}
-
-private enum FailingLLMClientError: Error {
-    case shouldNotBeCalled
-}
-
-private struct FailingLLMClient: LLMClient {
-    func nextAction(thread: ChatThread, userMessage: String, tools: [ToolDefinition]) async throws -> AgentAction {
-        throw FailingLLMClientError.shouldNotBeCalled
-    }
-}
-
-private struct FixedSayLLMClient: LLMClient {
-    var message: String
-
-    func nextAction(thread: ChatThread, userMessage: String, tools: [ToolDefinition]) async throws -> AgentAction {
-        .say(message)
-    }
-}
