+++ Tests/QuillCodeParityTests/ParityAgentGateTests.swift
@@ -101,19 +101,30 @@ final class ParityAgentGateTests: QuillCodeParityTestCase {
     }
 
     func testAgentBehaviorTestsUseFocusedSuites() throws {
-        let immediateTests = try Self.agentTestSourceText(named: "AgentImmediateActionTests.swift")
+        let immediateSupport = try Self.agentTestSourceText(named: "AgentImmediateActionTestSupport.swift")
+        let immediateShellTests = try Self.agentTestSourceText(named: "AgentImmediateShellActionTests.swift")
+        let immediateGitTests = try Self.agentTestSourceText(named: "AgentImmediateGitActionTests.swift")
+        let immediateFileTests = try Self.agentTestSourceText(named: "AgentImmediateFileActionTests.swift")
+        let immediateNegationTests = try Self.agentTestSourceText(named: "AgentImmediateNegationTests.swift")
         let toolLoopTests = try Self.agentTestSourceText(named: "AgentToolLoopTests.swift")
         let streamingTests = try Self.agentTestSourceText(named: "AgentStreamingTests.swift")
         let pullRequestTests = try Self.agentTestSourceText(named: "MockLLMClientPullRequestTests.swift")
         let finalAnswerTests = try Self.agentTestSourceText(named: "AgentFinalAnswerBuilderTests.swift")
         let supportTests = try Self.agentTestSourceText(named: "AgentTestSupport.swift")
 
-        XCTAssertTrue(immediateTests.contains("testRunWhoamiExecutesImmediately"), "Immediate command execution should live in focused agent action tests.")
+        XCTAssertTrue(immediateShellTests.contains("testRunWhoamiExecutesImmediately"), "Shell immediacy should live in focused shell action tests.")
+        XCTAssertTrue(immediateGitTests.contains("testCommitChangesExecutesImmediately"), "Git immediacy should live in focused git action tests.")
+        XCTAssertTrue(immediateFileTests.contains("testMakeHelloWorldFileExecutesImmediately"), "File immediacy should live in focused file action tests.")
+        XCTAssertTrue(immediateNegationTests.contains("testNegatedWhoamiDoesNotPreflight"), "Negated-action guardrails should live in focused negation tests.")
+        XCTAssertTrue(immediateSupport.contains("assertSingleSuccessfulToolResult"), "Immediate-action test helpers should be shared instead of copied.")
         XCTAssertTrue(toolLoopTests.contains("testAgentContinuesAcrossMultipleToolCallsInOneTurn"), "Tool loop behavior should live in focused agent tool-loop tests.")
         XCTAssertTrue(streamingTests.contains("testStreamingToolActionReportsStatusAndExecutes"), "Streaming behavior should live in focused agent streaming tests.")
         XCTAssertTrue(pullRequestTests.contains("testPullRequestMergeUsesStructuredToolCall"), "Mock PR planning should live in focused mock PR tests.")
         XCTAssertTrue(finalAnswerTests.contains("testBrowserInspectFinalAnswerSummarizesPage"), "Final-answer copy should live in focused final-answer tests.")
         XCTAssertTrue(supportTests.contains("struct FixedToolLLMClient"), "Agent test fakes should be shared instead of duplicated across focused suites.")
+        XCTAssertFalse(FileManager.default.fileExists(atPath: Self.packageRoot()
+            .appendingPathComponent("Tests/QuillCodeAgentTests/AgentImmediateActionTests.swift")
+            .path), "AgentImmediateActionTests.swift should not regrow as a broad mixed-behavior suite.")
         XCTAssertFalse(FileManager.default.fileExists(atPath: Self.packageRoot()
             .appendingPathComponent("Tests/QuillCodeAgentTests/AgentTests.swift")
             .path), "AgentTests.swift should not regrow as a broad mixed-behavior suite.")
